+++ src/config/format.rs
@@ -1,8 +1,22 @@
-use ratatui::prelude::Style;
+use crate::config::color_conversion::Style;
 use scraper::ElementRef;
-use std::collections::{HashMap, HashSet};
+use std::collections::{BTreeMap, HashMap, HashSet};
 
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone, Default, Eq, PartialEq, Hash)]
+pub enum TagData {
+    #[default]
+    None,
+    Styled(Style),
+}
+
+#[derive(Debug, Clone, Default, Eq, PartialEq, Hash)]
+pub struct TagDataIdentifier {
+    data: TagData,
+    classes: BTreeMap<String, TagData>,
+    ids: BTreeMap<String, TagData>,
+}
+
+#[derive(Debug, Clone, Default, Eq, PartialEq)]
 pub struct TagIdentifier {
     unconditional: bool,
     classes: HashSet<String>,
@@ -14,7 +28,7 @@ pub struct FormatConfig {
     pub ignored_tags: HashMap<String, TagIdentifier>,
     pub block_elements: HashMap<String, TagIdentifier>,
     pub indent_elements: HashMap<String, TagIdentifier>,
-    pub tag_styles: HashMap<String, Style>,
+    pub tag_styles: HashMap<String, TagDataIdentifier>,
 }
 
 impl FormatConfig {
@@ -27,11 +41,17 @@ impl FormatConfig {
         let ignored_tags_map = Self::build_map_from_selectors(ignored_tags);
         let block_elements_map = Self::build_map_from_selectors(block_elements);
         let indent_elements_map = Self::build_map_from_selectors(indent_elements);
+        let tag_styles_map = Self::build_data_map_from_selectors(
+            tag_styles
+                .into_iter()
+                .map(|(str, sty)| (str, TagData::Styled(sty)))
+                .collect(),
+        );
         Self {
             ignored_tags: ignored_tags_map,
             block_elements: block_elements_map,
             indent_elements: indent_elements_map,
-            tag_styles,
+            tag_styles: tag_styles_map,
         }
     }
 
@@ -59,6 +79,39 @@ impl FormatConfig {
                 .extend(classes.into_iter().map(String::from));
             if let Some(id) = id {
                 tag_identifier.ids.insert(id.to_string());
+            }
+        }
+        ignored_tags_map
+    }
+
+    fn build_data_map_from_selectors(
+        tag_to_data: HashSet<(String, TagData)>,
+    ) -> HashMap<String, TagDataIdentifier> {
+        let mut ignored_tags_map: HashMap<String, TagDataIdentifier> = HashMap::new();
+        for (tag, tag_data) in tag_to_data {
+            let mut class_split = tag.split('.');
+            let tag = class_split.next().unwrap_or_else(|| {
+                log::error!("Invalid css selector - must be of the format TAG#ID.CLASS, {tag}");
+                ""
+            });
+            let classes = class_split.collect::<Vec<&str>>();
+            let mut id_split = tag.split('#');
+            let tag = id_split.next().unwrap_or_else(|| {
+                log::error!("Invalid css selector - must be of the format TAG#ID.CLASS, {tag}");
+                ""
+            });
+            let id = id_split.next();
+            let tag_identifier = ignored_tags_map.entry(tag.to_string()).or_default();
+            if classes.is_empty() && id.is_none() {
+                tag_identifier.data = tag_data.clone();
+            }
+            for class in classes {
+                tag_identifier
+                    .classes
+                    .insert(class.to_string(), tag_data.clone());
+            }
+            if let Some(id) = id {
+                tag_identifier.ids.insert(id.to_string(), tag_data.clone());
             }
         }
         ignored_tags_map
@@ -110,7 +163,66 @@ impl FormatConfig {
         Self::matches_tag(element, tag_identifier) || Self::matches_tag(element, general_identifier)
     }
 
-    pub fn style_for_tag(&self, tag: &str) -> Option<&Style> {
-        self.tag_styles.get(tag)
+    pub fn style_for_tag(&self, element: &ElementRef) -> Option<Style> {
+        let general_identifier = self.tag_styles.get("");
+        let mut head_style = vec![];
+        let mut class_style = vec![];
+        let mut id_style = vec![];
+        if let Some(general) = general_identifier {
+            Self::insert_values(
+                element,
+                general,
+                &mut head_style,
+                &mut class_style,
+                &mut id_style,
+            );
+        }
+        let tag_identifier = self.tag_styles.get(element.value().name());
+        if let Some(tag) = tag_identifier {
+            Self::insert_values(
+                element,
+                tag,
+                &mut head_style,
+                &mut class_style,
+                &mut id_style,
+            );
+        }
+        let mut all_styles = vec![];
+        all_styles.extend(head_style);
+        all_styles.extend(class_style);
+        all_styles.extend(id_style);
+        if all_styles.is_empty() {
+            return None;
+        }
+        let mut merged_style = Style::default();
+        for style in all_styles {
+            merged_style = merged_style.patch(&style);
+        }
+        Some(merged_style)
+    }
+
+    fn insert_values(
+        element: &ElementRef,
+        identifier: &TagDataIdentifier,
+        head_style: &mut Vec<Style>,
+        class_style: &mut Vec<Style>,
+        id_style: &mut Vec<Style>,
+    ) {
+        if let TagData::Styled(style) = identifier.data.clone() {
+            head_style.push(style);
+        }
+        if let Some(classes) = element.value().attr("class") {
+            class_style.extend(classes.split_whitespace().filter_map(|class| {
+                identifier.classes.get(class).and_then(|td| match td {
+                    TagData::Styled(style) => Some(style),
+                    TagData::None => None,
+                })
+            }));
+        }
+        if let Some(id) = element.value().id() {
+            if let Some(TagData::Styled(style)) = identifier.ids.get(id) {
+                id_style.push(*style);
+            }
+        }
     }
 }