+++ src/config/site.rs
@@ -1,5 +1,5 @@
 use crate::cli::command::CacheMode;
-use crate::config::color_conversion::convert_styles;
+use crate::config::color_conversion::Style;
 use crate::config::files::config_location;
 use crate::config::format::FormatConfig;
 use crate::config::glob_generation::generate_globs;
@@ -86,6 +86,7 @@ impl SitePicker {
         ignored_additional: &[String],
         no_block: bool,
         cache_mode: Option<&CacheMode>,
+        styles: &[(String, Style)],
     ) -> Self {
         let mut site: SiteRawConfig = toml::from_str(DEFAULT_CONFIG)
             .map_err(|e| println!("{e}"))
@@ -94,7 +95,7 @@ impl SitePicker {
             .map(|u_config| override_defaults_site(&mut site, u_config));
 
         let base_site_config =
-            Self::create_base_site_config(&site, ignored_additional, no_block, cache_mode);
+            Self::create_base_site_config(&site, ignored_additional, no_block, cache_mode, styles);
         let mut site_mapping = HashMap::new();
         site_mapping.insert(String::new(), base_site_config);
         for (url, filenames) in custom_configs {
@@ -110,6 +111,7 @@ impl SitePicker {
                     ignored_additional,
                     no_block,
                     cache_mode,
+                    styles,
                 ),
             );
         }
@@ -126,8 +128,9 @@ impl SitePicker {
         ignored_additional: &[String],
         no_block: bool,
         cache_mode: Option<&CacheMode>,
+        styles: &[(String, Style)],
     ) -> SiteConfig {
-        let format = Self::create_format_config(raw, ignored_additional, no_block);
+        let format = Self::create_format_config(raw, ignored_additional, no_block, styles);
         let cache = Self::create_cache_config(cache_mode, raw);
         let syntax = Self::create_syntax_config(raw);
         let call = Self::create_call_config(raw);
@@ -166,6 +169,7 @@ impl SitePicker {
         config: &SiteRawConfig,
         ignored_additional: &[String],
         no_block: bool,
+        styles: &[(String, Style)],
     ) -> FormatConfig {
         let mut ignored_tags: HashSet<String> = config
             .format
@@ -187,8 +191,14 @@ impl SitePicker {
             .as_ref()
             .map(|format| format.indent_elements.iter().cloned().collect())
             .unwrap_or_default();
-        let tag_styles = convert_styles(&config.styles);
-        FormatConfig::new(ignored_tags, block_elements, indent_elements, tag_styles)
+        let mut existing_styles = config.styles.clone();
+        existing_styles.extend(styles.iter().cloned());
+        FormatConfig::new(
+            ignored_tags,
+            block_elements,
+            indent_elements,
+            existing_styles.clone(),
+        )
     }
 
     fn create_cache_config(cache_mode: Option<&CacheMode>, config: &SiteRawConfig) -> CacheConfig {